requirements.txt
langgraph
langchain
langchain-openai
langchain-tavily
python-dotenv
numpy
scikit-learn
python-dateutil
pandas
streamlit